Make builder generator optional.
